client/gtk2: Add IBUS_AUTO_PREEDIT_COMMIT_APPS env for Hangul
Firefox no longer connect "button-press-event" but uses
GtkGestureMultiPress in GtkWindowPrivate and connects "pressed"
instead and do not propagate the event with "button-press-event".

We don't commit the preedit text in the client here for the
workaround and let firefox commit the preedit text to avoid the
double preedit text with mouse click.

BUG=https://github.com/ibus/ibus/issues/1980