bus: Add bus_component_check for GitHub action + Docker
Currently g_spawn_async() is failed in GitHub action + Docker, which
is called by bus_component_start() because of no permission to close
file descriptors of the child processes in g_spawn_async().
Add G_SPAWN_LEAVE_DESCRIPTORS_OPEN to avoid the failre of
g_spawn_async() for the workaround.