ui/gtk3: Enable libdbusmenu in Indicator
IBus panel now uses libdbusmenu to settle the panel menu coordiantes
in Plasma Wayland desktop session since the Wayland display is required
to use the Wayland input-method protocol.

Now g_bus_get (G_BUS_TYPE_SESSION) is moved to Indicator to follow
the AppIndicator implementation strictly and a new API
ibus_service_class_free_interfaces() is added in libibus.so in case
IBusService constructor does not need GDBusConnection.

Previously GtkMenu absolute cooridiantes are used with Xorg display
and AppIndicator "Activate" and "ContextMenu" D-Bus methods are used
but now the coordiates are transferred to libdbusmenu and "Menu" D-Bus
method is available and the activate menu only is shown by default.

The workaround is to use the middle mouse click to toggle the activate
and context menus at the moment.

BUG=https://github.com/ibus/ibus/issues/2408